Revert "create two distinct classes TravelCosts (car, walk, bicycle) and PublicTravelCosts (all public transport modes) + create a GTFS class to handle download / aggregation / GTFS router object preparation"
This reverts commit ebed714206d851630afd8c4ea7e8f8dc94d70d19.
