+++ mobility/__init__.py
@@ -2,4 +2,4 @@ from .set_params import set_params
 from .trip_sampler import TripSampler
 from .transport_zones import TransportZones
 from .travel_costs import TravelCosts
-from .public_transport_travel_costs import PublicTransportTravelCosts
+from .gtfs import GTFS
